Added files copied from Linux v5.4
These files all carry SPDX GPL2+ indicators.  Based on the information
available to me, I have a good-faith belief that they can be combined
into this AGPLv3 project and licensed AGPLv3.

This commit only touches t/tree.  Therefore, if the test suite later
needs to be changed, it will be sufficient to revert this commit.

The files used are those involved in the x64 compilation of
drivers/i2c/i2c-core-base.c, filtered to remove any that aren't
tagged GPL2+.

Note that t/tree itself is not a Git repo or submodule.  It is a
collection of files that test scripts can use to populate a Git repo.