update.py: Exit if there is no tag to add
This prevent the script from locking at line 578 because UpdateIds will never send an event to wake up the threads (line 91) if there is no tag to add.

Also change len(tag_buf) to num_tags in line 548, no need to compute the
length again as we already have it.

Signed-off-by: Maxime Chretien <maxime.chretien@bootlin.com>