Prefer GAS Lexer for .S files
Currently, Pygments often picks SLexer for .S files. It's a lexer for
files related to the R language.

https://pygments.org/docs/lexers/#pygments.lexers.r.SLexer

```
>>> import pygments.lexers
>>> pygments.lexers.guess_lexer_for_filename('arch/x86/boot/header.S',
        open('arch/x86/boot/header.S').read())
<pygments.lexers.SLexer>
```

This commits makes sure Elixir prefers the GAS lexer for .S files
instead of SLexer.
It's not the best heuristic, but better than picking the SLexer class
sometimes.
Currently, Elixir does not index any R language projects. In the future,
it would be better to specify per-project information about what
languages are used.