web: Explicitly close Query after finishing request
API closes Query and isn't known to leak memory.
BsdDB seems to have automatic destructors, but maybe there are
some caveats.
The documentation does not say that objects have to be closed
explicitly... But it probably won't hurt. And this is the only
hint I found when trying to debug the leak so far.

https://hg.jcea.es/pybsddb/file/6.2.9/Modules/_bsddb.c#l8943
https://pybsddb.sourceforge.net/bsddb3.html