web: Remove whitespace from versions list
Due to redundant whitespace, HTML of versions list with all Linux
tags (including tags that are normally filtered away) takes around
2 megabytes.

This commit aims to reduce that number by removing all whitespace from
the sidebar.
This cut the size of the versions list HTML to around 300 kilobytes.

Large sidebar HTML is a problem - the version list on the official
instance is so large, that the subtrees for versions v2-v0 aren't
visible immediately after opening the page. Since the sidebar is
now above content, the content does not load until the sidebar loads.