script/query: don't restrict `./script.sh get-latest` to a single tag
`./script.sh get-latest <offset>` gets the full list of tags, filters
it, sorts it then returns a single result. On the Python side, it gets
the first one. If that works, it uses it, else it tries the second one,
etc.

That is a weird implementation: modify get-latest to return all tags so
that Python code only has to spawn a single subprocess.

Also, rename it from `get-latest` to `get-latest-tags`. This makes
things more explicit (what latest?) and also explicits that more than
one tag is required.

Also, argument 1 is supposed to be an offset.
No custom implementation of get_latest() did implement that.

Signed-off-by: Théo Lebrun <theo.lebrun@bootlin.com>