Dockerfile: set environment variable $ELIXIR_ROOT
Set env variable $ELIXIR_ROOT inside the Dockerfile. That is used by
utils/index-all-repositories to know where to put repo/data directories.

With that adjustment the docker setup is working as indicated in the
README.

Fix the following error:

$ docker exec -it -e PYTHONUNBUFFERED=1 elixir-container /bin/bash -c \
  'export "PATH=/usr/local/elixir/venv/bin:$PATH" ; /usr/local/elixir/utils/index-all-repositories'
Error: ELIXIR_ROOT environment variable not set
It's where Elixir data are stored