treewide: add guardrails to ensure definitions are cached
Add a list (CACHED_DEFINITIONS_FAMILIES) that tells us which families
have their definitions cached. We use that to assert at DB.__init__()
and q.query('file') that everything is working as expected.

If someone modifies lib.getFileFamily() for example, we'll get an
explicit warning that we should add a definitions cache to that new
family.

Signed-off-by: Théo Lebrun <theo.lebrun@bootlin.com>