utils/index: add lock on indexing process
Logs indicate that an indexing took longer than twelve hours. That
implies a second `index` call called. As such, it isn't a massive deal.
BUT it might be the reason the LLVM database got corrupted. The
indexing container cannot be started twice has it always has the same
name so Docker fails `docker run` if it already is running. A
corruption sequence could be something like:
 - Indexing 1, duplicate database from A to B
 - Indexing 1, start indexing on B
 - Indexing 2, duplicate database from A to B
=> Database files are copied on top of files that are being worked on.

Use flock (rather than a PID file technique) to protect the update_index
call. If lock is taken, the call fails and the script stops early
because we crash on non-zero returns using `set -e`.

Signed-off-by: Théo Lebrun <theo.lebrun@bootlin.com>